requirements.txt
# ===== CORE TRADING LIBRARIES =====
alpaca-trade-api==3.0.2
yfinance==0.2.22

# ===== AI/ML LIBRARIES =====
anthropic==0.7.0

# ===== WEB FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ===== DATABASE =====
sqlalchemy==2.0.23
alembic==1.12.1

# ===== DATA PROCESSING =====
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# ===== ASYNC SUPPORT =====
# Note: aiohttp version constrained by alpaca-trade-api
aiohttp==3.8.2
aiofiles==23.2.0

# ===== UTILITIES =====
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pydantic==2.5.0
python-dateutil==2.8.2

# ===== DEVELOPMENT TOOLS =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# ===== MONITORING =====
prometheus-client==0.19.0
structlog==23.2.0

# ===== TECHNICAL ANALYSIS =====
# Note: TA-Lib requires special installation on macOS - see installation instructions
# Alternative: pandas-ta (easier to install)
pandas-ta==0.3.14b